chore: move "client" stuff to metastoreproto
Fixes #1128
