+++ crates/metastore/src/database.rs
@@ -1,9 +1,6 @@
 //! Module for handling the catalog for a single database.
 use crate::errors::{MetastoreError, Result};
 use crate::storage::persist::Storage;
-use crate::validation::{
-    validate_database_tunnel_support, validate_object_name, validate_table_tunnel_support,
-};
 use metastoreproto::types::catalog::{
     CatalogEntry, CatalogState, CredentialsEntry, DatabaseEntry, EntryMeta, EntryType,
     FunctionEntry, FunctionType, SchemaEntry, TableEntry, TunnelEntry, ViewEntry,
@@ -13,6 +10,9 @@ use metastoreproto::types::options::{
 };
 use metastoreproto::types::service::Mutation;
 use metastoreproto::types::storage::{ExtraState, PersistedCatalog};
+use metastoreproto::validation::{
+    validate_database_tunnel_support, validate_object_name, validate_table_tunnel_support,
+};
 use once_cell::sync::Lazy;
 use pgrepr::oid::FIRST_AVAILABLE_ID;
 use sqlbuiltins::builtins::{
