+++ crates/metastore/src/errors.rs
@@ -1,8 +1,10 @@
-use tonic::metadata::AsciiMetadataValue;
-use tracing::warn;
+use metastoreproto::errors::{ResolveErrorStrategy, RESOLVE_ERROR_STRATEGY_META};
 
 #[derive(thiserror::Error, Debug)]
 pub enum MetastoreError {
+    #[error("Catalog version mismatch; have: {have}, need: {need}")]
+    VersionMismtatch { have: u64, need: u64 },
+
     #[error("Duplicate name: {0}")]
     DuplicateName(String),
 
@@ -42,26 +44,11 @@ pub enum MetastoreError {
     #[error("Missing entry: {0}")]
     MissingEntry(u32),
 
-    #[error("Tunnel '{tunnel}' not supported by datasource '{datasource}'")]
-    TunnelNotSupportedByDatasource { tunnel: String, datasource: String },
-
-    #[error("Credentials '{credentials}' not supported by datasource '{datasource}'")]
-    CredentialsNotSupportedByDatasource {
-        credentials: String,
-        datasource: String,
-    },
-
-    #[error("Format '{format}' not supported by datasource '{datasource}'")]
-    FormatNotSupportedByDatasource { format: String, datasource: String },
-
     #[error("Tunnel '{tunnel} not supported for {action}'")]
     TunnelNotSupportedForAction {
         tunnel: String,
         action: &'static str,
     },
-
-    #[error("Catalog version mismatch; have: {have}, need: {need}")]
-    VersionMismtatch { have: u64, need: u64 },
 
     #[error("Invalid database id: {0:?}")]
     InvalidDatabaseId(Vec<u8>),
@@ -101,8 +88,8 @@ pub enum MetastoreError {
     #[error(transparent)]
     ObjectStore(#[from] object_store::Error),
 
-    #[error("Feature unimplemented: {0}")]
-    Unimplemented(&'static str),
+    #[error(transparent)]
+    MetastoreProto(#[from] metastoreproto::errors::MetastoreProtoError),
 }
 
 pub type Result<T, E = MetastoreError> = std::result::Result<T, E>;
@@ -118,71 +105,11 @@ impl From<MetastoreError> for tonic::Status {
     }
 }
 
-pub const RESOLVE_ERROR_STRATEGY_META: &str = "resolve-error-strategy";
-
-/// Additional metadata to provide a hint to the client on what it can do to
-/// automatically resolve an error.
-///
-/// These are only suggestions, and correctness should not depend on the action
-/// of the client.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ResolveErrorStrategy {
-    /// Client may refetch the latest catalog and retry the operation.
-    FetchCatalogAndRetry,
-
-    /// No known steps to resolve the error. Client should bubble the error up.
-    Unknown,
-}
-
-impl ResolveErrorStrategy {
-    pub fn to_metadata_value(&self) -> AsciiMetadataValue {
-        match self {
-            Self::Unknown => AsciiMetadataValue::from_static("0"),
-            Self::FetchCatalogAndRetry => AsciiMetadataValue::from_static("1"),
-        }
-    }
-
-    pub fn from_bytes(bs: &[u8]) -> ResolveErrorStrategy {
-        match bs.len() {
-            1 => match bs[0] {
-                b'0' => Self::Unknown,
-                b'1' => Self::FetchCatalogAndRetry,
-                _ => Self::parse_err(bs),
-            },
-            _ => Self::parse_err(bs),
-        }
-    }
-
-    fn parse_err(bs: &[u8]) -> ResolveErrorStrategy {
-        warn!(?bs, "failed getting resolve strategy from bytes");
-        ResolveErrorStrategy::Unknown
-    }
-}
-
 impl MetastoreError {
     pub fn resolve_error_strategy(&self) -> ResolveErrorStrategy {
         match self {
             Self::VersionMismtatch { .. } => ResolveErrorStrategy::FetchCatalogAndRetry,
             _ => ResolveErrorStrategy::Unknown,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn strategy_roundtrip() {
-        let variants = [
-            ResolveErrorStrategy::FetchCatalogAndRetry,
-            ResolveErrorStrategy::Unknown,
-        ];
-
-        for strat in variants {
-            let val = strat.to_metadata_value();
-            let out = ResolveErrorStrategy::from_bytes(val.as_bytes());
-            assert_eq!(strat, out);
         }
     }
 }
