+++ crates/metastore/src/lib.rs
@@ -2,8 +2,6 @@
 pub mod errors;
 pub mod local;
 pub mod srv;
-pub mod types;
-pub mod validation;
 
 mod database;
 mod storage;
