+++ crates/metastoreproto/src/types/options.rs
@@ -1097,3 +1097,99 @@ impl From<CredentialsOptionsAws> for options::CredentialsOptionsAws {
         }
     }
 }
+
+#[derive(Debug, Clone)]
+pub enum CopyToDestinationOptions {
+    Local(CopyToDestinationOptionsLocal),
+    Gcs(CopyToDestinationOptionsGcs),
+    S3(CopyToDestinationOptionsS3),
+}
+
+impl CopyToDestinationOptions {
+    pub const LOCAL: &str = "local";
+    pub const GCS: &str = "gcs";
+    pub const S3_STORAGE: &str = "s3";
+
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Self::Local(_) => Self::LOCAL,
+            Self::Gcs(_) => Self::GCS,
+            Self::S3(_) => Self::S3_STORAGE,
+        }
+    }
+
+    pub fn location(&self) -> &str {
+        match self {
+            Self::Local(CopyToDestinationOptionsLocal { location }) => location,
+            Self::Gcs(CopyToDestinationOptionsGcs { location, .. }) => location,
+            Self::S3(CopyToDestinationOptionsS3 { location, .. }) => location,
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct CopyToDestinationOptionsLocal {
+    pub location: String,
+}
+
+#[derive(Debug, Clone)]
+pub struct CopyToDestinationOptionsGcs {
+    pub service_account_key: Option<String>,
+    pub bucket: String,
+    pub location: String,
+}
+
+#[derive(Debug, Clone)]
+pub struct CopyToDestinationOptionsS3 {
+    pub access_key_id: Option<String>,
+    pub secret_access_key: Option<String>,
+    pub region: String,
+    pub bucket: String,
+    pub location: String,
+}
+
+#[derive(Debug, Clone)]
+pub enum CopyToFormatOptions {
+    Csv(CopyToFormatOptionsCsv),
+    Parquet(CopyToFormatOptionsParquet),
+    Json(CopyToFormatOptionsJson),
+}
+
+impl Default for CopyToFormatOptions {
+    fn default() -> Self {
+        Self::Csv(CopyToFormatOptionsCsv {
+            delim: b',',
+            header: true,
+        })
+    }
+}
+
+impl CopyToFormatOptions {
+    pub const CSV: &str = "csv";
+    pub const PARQUET: &str = "parquet";
+    pub const JSON: &str = "json";
+
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            Self::Csv(_) => Self::CSV,
+            Self::Parquet(_) => Self::PARQUET,
+            Self::Json(_) => Self::JSON,
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct CopyToFormatOptionsCsv {
+    pub delim: u8,
+    pub header: bool,
+}
+
+#[derive(Debug, Clone)]
+pub struct CopyToFormatOptionsParquet {
+    pub row_group_size: usize,
+}
+
+#[derive(Debug, Clone)]
+pub struct CopyToFormatOptionsJson {
+    pub array: bool,
+}
