+++ crates/sqlexec/Cargo.toml
@@ -8,7 +8,6 @@ edition = {workspace = true}
 [dependencies]
 logutil = {path = "../logutil"}
 pgrepr = {path = "../pgrepr"}
-metastore = {path = "../metastore"}
 metastoreproto = { path = "../metastoreproto" }
 datafusion_planner = { path = "../datafusion_planner" }
 telemetry = {path = "../telemetry"}
@@ -34,3 +33,4 @@ url.workspace = true
 [dev-dependencies]
 tempfile = "3"
 tower = "0.4"
+metastore = { path = "../metastore" }
