+++ crates/sqlexec/src/planner/logical_plan.rs
@@ -5,7 +5,7 @@ use datafusion::datasource::TableProvider;
 use datafusion::logical_expr::{Explain, LogicalPlan as DfLogicalPlan};
 use datafusion::scalar::ScalarValue;
 use datafusion::sql::sqlparser::ast;
-use metastore::types::{CopyToDestinationOptions, CopyToFormatOptions};
+use metastoreproto::types::options::{CopyToDestinationOptions, CopyToFormatOptions};
 use metastoreproto::types::options::{
     CredentialsOptions, DatabaseOptions, TableOptions, TableOptionsInternal, TunnelOptions,
 };
