+++ crates/sqlexec/src/session.rs
@@ -22,8 +22,8 @@ use datasources::object_store::gcs::GcsTableAccess;
 use datasources::object_store::local::LocalTableAccess;
 use datasources::object_store::s3::S3TableAccess;
 use futures::StreamExt;
-use metastore::types::{CopyToDestinationOptions, CopyToFormatOptions};
 use metastoreproto::session::SessionCatalog;
+use metastoreproto::types::options::{CopyToDestinationOptions, CopyToFormatOptions};
 use object_store::path::Path as ObjectStorePath;
 use object_store::ObjectStore;
 use pgrepr::format::Format;
